requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
SQLAlchemy>=2.0.34
Authlib==1.2.1
python-dotenv==1.0.0
requests==2.31.0
eventlet==0.33.3
gunicorn==21.2.0
websocket-client==1.8.0
pylast==5.2.0
